--- app/src/BlocksPage.php
@@ -1,8 +1,0 @@
-<?php
-
-class BlocksPage extends Page
-{
-    private static $table_name = 'BlocksPage';
-
-    private static $description = 'A modular page composed of content blocks';
-}
