+++ app/src/Page.php
@@ -12,28 +12,6 @@ namespace {
 
     class Page extends SiteTree
     {
-        private static $extensions = [
-            ElementalPageExtension::class,
-        ];
-
-        private static $has_one = [
-            'HeaderElements' => ElementalArea::class,
-        ];
-
-        public function getCMSFields()
-        {
-            $fields = parent::getCMSFields();
-            $headerArea = $fields->fieldByName('Root.Main.HeaderElements');
-            $fields->removeByName('Root.Main.HeaderElements');
-            $fields->addFieldToTab('Root.Main', $headerArea, 'ElementalArea');
-            $fields->addFieldToTab('Root.Main', HeaderField::create('HeaderBlocksHeader', 'Header blocks'), 'HeaderElements');
-            $fields->addFieldToTab('Root.Main', HeaderField::create('ContentBlocksHeader', 'Content blocks'), 'ElementalArea');
-//            $fields->fieldByName('Root.Main.HeaderElements')->setTypes([
-//                CarouselBlock::class,
-//                BannerBlock::class,
-//            ]);
-            return $fields;
-        }
 
     }
 }
