--- app/src/Pages/HomePageController.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace SilverStripe\Bambusa\Pages;
-
-use PageController;
-
-class HomePageController extends PageController
-{
-
-}
