+++ lang/lang/zh.yml
@@ -1,0 +1,4 @@
+zh:
+  SilverStripe\Bambusa\Blocks\CarouselItem:
+    PLURALNAME: 旋转项
+    SINGULARNAME: 旋转项
